rearrange & add skpca option
